requirements.txt
fastapi
uvicorn
pymongo
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-dotenv